requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
bcrypt
PyJWT
